refactor: refactor test helpers
