+++ tests/clipboard/cut.ts
@@ -1,5 +1,5 @@
 import userEvent from '#src'
-import {setup} from '#testHelpers/utils'
+import {setup} from '#testHelpers'
 
 test('cut selected value', async () => {
   const {element, getEvents} = setup<HTMLInputElement>(
