+++ tests/clipboard/paste.ts
@@ -1,5 +1,5 @@
 import userEvent from '#src'
-import {setup} from '#testHelpers/utils'
+import {setup} from '#testHelpers'
 
 test('should paste text in input', async () => {
   const {element, getEventSnapshot} = setup('<input />')
