+++ tests/convenience/hover.ts
@@ -1,5 +1,5 @@
 import userEvent, {PointerEventsCheckLevel} from '#src'
-import {setup} from '#testHelpers/utils'
+import {setup} from '#testHelpers'
 
 describe.each([
   ['hover', {events: ['over', 'enter', 'move']}],
