+++ tests/event/behavior/keydown.ts
@@ -1,5 +1,5 @@
 import userEvent from '#src'
-import {setup} from '#testHelpers/utils'
+import {setup} from '#testHelpers'
 
 test.each(['Backspace', 'Delete', 'End', 'Home', 'PageUp', 'PageDown'])(
   'implement no keydown behavior for [%s] outside of editable context',
