+++ tests/event/behavior/keyup.ts
@@ -1,5 +1,5 @@
 import userEvent from '#src'
-import {setup} from '#testHelpers/utils'
+import {setup} from '#testHelpers'
 
 describe('release [Space]', () => {
   test.each([
