+++ tests/utils/edit/calculateNewValue.ts
@@ -1,5 +1,5 @@
 import userEvent from '#src'
-import {setup} from '#testHelpers/utils'
+import {setup} from '#testHelpers'
 
 // TODO: focus the maxlength tests on the tested aspects
